Accidently merged unsigned-char-to-double frequency patch. Reversing that.
git-svn-id: svn://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx_exp@5421 ab8295b8-cf94-4d9e-aec4-7959e3be5d79

+++ Modulator.cpp
@@ -43,7 +43,7 @@ Modulator::Modulator (unsigned frameRate, unsigned periodLengthInSeconds,
 }
 
 void Modulator::start (unsigned symbolsLength, double framesPerSymbol,
-                       double frequency, double toneSpacing,
+                       unsigned frequency, double toneSpacing,
                        SoundOutput * stream, Channel channel,
                        bool synchronize, double dBSNR)
 {
